refactor: move componenent to origin reposaability
