+++ src/components/containers/Home/HomeModalContent/Techs/index.tsx
