+++ src/components/containers/Home/HomeModalContent/Techs/styles.ts
