+++ src/components/containers/Home/HomeModalContent/index.tsx
@@ -1,4 +1,4 @@
-import Techs from '../../global/Techs'
+import Techs from './Techs'
 
 import { Container } from './styles'
