+++ src/components/containers/Home/HomeModalContent/styles.ts
