+++ src/components/containers/Home/Typewriter/index.tsx
