+++ src/components/containers/Home/index.tsx
@@ -2,8 +2,8 @@ import React, { useRef } from 'react'
 
 import Button from '../../global/Button'
 import Profile from '../../global/Profile'
-import Typewriter from '../../global/Typewriter'
-import HomeModalContent from '../HomeModalContent'
+import Typewriter from './Typewriter'
+import HomeModalContent from './HomeModalContent'
 import SocialMidia from '../../global/SocialMidia'
 import Modal, { IHandleVisibleOption, IModalHandles } from '../../global/Modal'
 
@@ -28,9 +28,11 @@ const Home: React.FC = () => {
   return (
     <Container>
       <Profile />
+
       <ContainerTypewriter>
         <Typewriter />
       </ContainerTypewriter>
+
       <ContainerButton>
         <Button
           label="About me"
@@ -39,7 +41,9 @@ const Home: React.FC = () => {
         <span>or</span>
         <Button label="Curriculum" onClick={handleOpenCurriculum} />
       </ContainerButton>
+
       <Modal ref={modalRef} modalContent={HomeModalContent} />
+
       <ContainerSocielMidia>
         <SocialMidia />
       </ContainerSocielMidia>
